Fixing to accept large numbers.

+++ eclipse workspace/OSGi Receita Exporter/src/main/java/com/edpichler/receita/interfaces/IExportacao.java
@@ -30,11 +30,11 @@ public interface IExportacao {
 	/**
 	 * Tamanho = 12
 	 */
-	public int getNumeroRegistroExportacao();
+	public long getNumeroRegistroExportacao();
 
 	/**
 	 * Tamanho = 12
 	 */
-	public int getNumeroDespachoExportacao();
+	public long getNumeroDespachoExportacao();
 
 }
